[wip] regroup the mpi primitives
