+++ src/impl/KokkosComm_isend.hpp
@@ -20,15 +20,12 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "KokkosComm_base.hpp"
+#include "KokkosComm_mpi.hpp"
 #include "KokkosComm_concepts.hpp"
 #include "KokkosComm_pack_traits.hpp"
 #include "KokkosComm_request.hpp"
 #include "KokkosComm_traits.hpp"
 #include "KokkosComm_comm_mode.hpp"
-
-// impl
-#include "KokkosComm_include_mpi.hpp"
 
 namespace KokkosComm::Impl {
