+++ src/impl/KokkosComm_recv.hpp
@@ -18,13 +18,10 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "KokkosComm_base.hpp"
+#include "KokkosComm_mpi.hpp"
 #include "KokkosComm_concepts.hpp"
 #include "KokkosComm_pack_traits.hpp"
 #include "KokkosComm_traits.hpp"
-
-// impl
-#include "KokkosComm_include_mpi.hpp"
 
 namespace KokkosComm::Impl {
 template <KokkosExecutionSpace ExecSpace, KokkosView RecvView>
